backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
passlib==1.7.4
pydantic==2.5.0
python-dotenv==1.0.0
pytest>=8.2